+++ src/01-document/index.js
@@ -1,3 +1,4 @@
+const Document = require('./Document');
 const redirects = require('./redirects');
 const disambig = require('./disambig');
 const preProcess = require('./preProcess');
@@ -22,7 +23,7 @@ const main = function(wiki, options) {
   if (redirects.isRedirect(wiki) === true) {
     data.type = 'redirect';
     data.redirectTo = redirects.parse(wiki);
-    return data;
+    return new Document(data, options);
   }
   //detect if page is just disambiguator page, and return
   if (disambig.isDisambig(wiki) === true) {
@@ -40,8 +41,8 @@ const main = function(wiki, options) {
   wiki = parse.categories(data, wiki);
   //parse all the headings, and their texts/sentences
   data.sections = parse.section(wiki, options) || [];
-
-  return data;
+  //all together now
+  return new Document(data, options);
 };
 
 module.exports = main;