+++ tests/page.test.js
@@ -59,7 +59,7 @@ test('statoil', t => {
   // (doc.text.Intro.length >= 1).should.be.true;
   t.equal(doc.categories().length, 4, 'cat-length');
   t.equal(doc.images().length, 1, 'img-length');
-  t.equal(doc.images(0).file, 'Fil:Statoil-Estonia.jpg', '');
+  t.equal(doc.images(0).file(), 'Fil:Statoil-Estonia.jpg', '');
   t.equal(doc.images(0).url(), 'https://upload.wikimedia.org/wikipedia/commons/8/87/Statoil-Estonia.jpg', t);
   t.end();
 });
@@ -70,7 +70,7 @@ test('raith rovers', t => {
   t.equal(doc.infoboxes(0).data.clubname.text(), 'Raith Rovers', '');
   t.equal(doc.categories().length, 10, 'cat-length');
   t.equal(doc.images().length, 2, 'img-length');
-  t.equal(doc.images(1).file, 'File:Stark\'s Park - geograph.org.uk - 204446.jpg', 'img-file');
+  t.equal(doc.images(1).file(), 'File:Stark\'s Park - geograph.org.uk - 204446.jpg', 'img-file');
   t.equal(
     doc.images(1).url(),
     'https://upload.wikimedia.org/wikipedia/commons/3/38/Stark\'s_Park_-_geograph.org.uk_-_204446.jpg',