+++ tests/template.test.js
@@ -220,8 +220,8 @@ The types of boats provided at each station and the launching methods vary depen
 hello there
 `;
   var doc = wtf(str);
-  t.equal(doc.templates().length, 5, 'got several templates');
-  t.equal(doc.templates('citation').length, 1, 'got citation template');
+  t.equal(doc.templates().length, 4, 'got several templates');
+  t.equal(doc.citations().length, 1, 'got citation template');
   t.equal(doc.templates('main').length, 1, 'got main template');
   t.equal(doc.templates('tracklist').length, 1, 'got tracklist template');
   t.equal(doc.templates('infobox').length, 1, 'got infobox template');